Fix missing debugging information
When building a static libjpeg for 8 bits, functions for 12 and 16 bits
are put in the library as well (for whatever reasons).  However, for
RelWithDebInfo builds, these objects store the debug info in separate
PDBs, but we don't ship these PDBs.  But even if we did, PHP's build
system wouldn't properly handle them, since it expects the basenames
of the .lib and the .pdb to match.

To work around this, we adjust the names of the PDBs in the first
place.